Changes in CountViewModel

+++ app/src/main/java/com/lumisdinos/measureandcount/ui/screens/count/CountEffect.kt
@@ -8,7 +8,7 @@ sealed class CountEffect {
     data class ShowSelectNotFoundToFindAreaConfirmationDialog(val chipboard: ChipboardUi) : CountEffect()
     data class ShowRemoveNotFoundFromFindAreaConfirmationDialog(val chipboard: ChipboardUi) : CountEffect()
     data class ShowSelectUnknownToFindAreaConfirmationDialog(val chipboard: ChipboardUi) : CountEffect()
-    data class ShowNotExceedingTargetQuantityDialog(val targetQuantity: Short, val enteredQuantity: Short) : CountEffect()
+    data class ShowNotExceedingTargetQuantityDialog(val targetQuantity: Int, val enteredQuantity: Int) : CountEffect()
     data class ShowWhatIsDialog(val questionType: QuestionType) : CountEffect()
     data class ShowSnackbar(val message: String) : CountEffect()
     data object ShowFieldDisabled : CountEffect()
